Refactored EdgelDetector. Moved all finding process into ARMarkerDetector.

+++ Augmented Reality/FrameDecorator.cpp
@@ -8,7 +8,8 @@ FrameDecorator::FrameDecorator(const size_t borderSize, const size_t regionSize,
 void FrameDecorator::drawRegionLines(cv::Mat& frame) {
 	for (size_t x = BORDER_SIZE; x < frame.cols; x += REGION_SIZE) {
 		for (size_t y = BORDER_SIZE; y < frame.rows; y += REGION_SIZE) {
-			cv::Rect region(x, y, x + REGION_SIZE - 1, y + REGION_SIZE - 1);
+			cv::Rect region(static_cast<int>(x), static_cast<int>(y), 
+				static_cast<int>(x + REGION_SIZE - 1), static_cast<int>(y + REGION_SIZE - 1));
 
 			cv::Scalar greenColor(0, 128, 0);
 
@@ -22,8 +23,9 @@ void FrameDecorator::drawSubRegionLines(cv::Mat& frame) {
 		for (size_t regionTop = BORDER_SIZE; regionTop < frame.rows; regionTop += REGION_SIZE) {
 			for (size_t x = 0; x < REGION_SIZE; x += STEP_SIZE) {
 				for (size_t y = 0; y < REGION_SIZE; y += STEP_SIZE) {
-					cv::Rect subregion(regionLeft + x, regionTop + y, 
-						regionLeft + x + STEP_SIZE - 1, regionTop + y + STEP_SIZE - 1);
+					cv::Rect subregion(static_cast<int>(regionLeft + x), static_cast<int>(regionTop + y),
+						static_cast<int>(regionLeft + x + STEP_SIZE - 1), 
+						static_cast<int>(regionTop + y + STEP_SIZE - 1));
 
 					cv::Scalar greenColor(128, 128, 128);
 
@@ -34,7 +36,7 @@ void FrameDecorator::drawSubRegionLines(cv::Mat& frame) {
 	}
 }
 
-void FrameDecorator::drawEdgels(cv::Mat & frame, std::vector<Edgel> edgels) {
+void FrameDecorator::drawEdgels(cv::Mat& frame, std::vector<Edgel> edgels) {
 	for (auto& edgel : edgels) {
 		drawEdgel(frame, edgel);
 	}
